Seperate event handling from the superpower handler
